Renamed api app to core
