+++ core/algorithms.py
