+++ core/queries.py
